+++ clang/docs/ReleaseNotes.rst
@@ -158,6 +158,10 @@ clang-format
 
 - ...
 
+- Option ``SpacesInLineCommentPrefix`` has been added to control the
+  number of spaces in a line comments prefix.
+
+
 libclang
 --------